requirements.txt
discord.py>=2.3.0
anthropic>=0.90.0
python-dotenv>=1.0.0
